+++ app/src/main/java/app/passwordstore/util/autofill/Api30AutofillResponseBuilder.kt
@@ -13,7 +13,9 @@ import android.service.autofill.FillCallback
 import android.service.autofill.FillRequest
 import android.service.autofill.FillResponse
 import android.service.autofill.Presentations
+import android.service.autofill.RegexValidator
 import android.service.autofill.SaveInfo
+import android.view.autofill.AutofillId
 import android.view.inputmethod.InlineSuggestionsRequest
 import android.widget.inline.InlinePresentationSpec
 import androidx.annotation.RequiresApi
@@ -21,11 +23,14 @@ import app.passwordstore.ui.autofill.AutofillDecryptActivity
 import app.passwordstore.ui.autofill.AutofillFilterView
 import app.passwordstore.ui.autofill.AutofillPublisherChangedActivity
 import app.passwordstore.ui.autofill.AutofillSaveActivity
+import app.passwordstore.util.extensions.sharedPrefs
+import app.passwordstore.util.settings.PreferenceKeys
 import com.github.androidpasswordstore.autofillparser.AutofillAction
 import com.github.androidpasswordstore.autofillparser.FillableForm
 import com.github.androidpasswordstore.autofillparser.fillWith
 import com.github.michaelbull.result.fold
 import java.io.File
+import java.util.regex.Pattern
 import logcat.LogPriority.ERROR
 import logcat.asLog
 import logcat.logcat
@@ -247,7 +252,9 @@ class Api30AutofillResponseBuilder private constructor(form: FillableForm) :
           makeHeaderMetadata(formOrigin.getPrettyIdentifier(context, untrusted = true)),
         )
       )
-      makeSaveInfo()?.let { setSaveInfo(it) }
+      makeSaveInfo(context)?.let {
+        setSaveInfo(it)
+      }
       setClientState(clientState)
       setIgnoredIds(*ignoredIds.toTypedArray())
       build()
@@ -257,8 +264,12 @@ class Api30AutofillResponseBuilder private constructor(form: FillableForm) :
   // TODO: Support multi-step authentication flows in apps via FLAG_DELAY_SAVE
   // See:
   // https://developer.android.com/reference/android/service/autofill/SaveInfo#FLAG_DELAY_SAVE
-  private fun makeSaveInfo(): SaveInfo? {
-    if (!canBeSaved) return null
+  private fun makeSaveInfo(context: Context): SaveInfo? {
+    if (
+      !canBeSaved ||
+        !context.sharedPrefs.getBoolean(PreferenceKeys.AUTOFILL_ASK_TO_SAVE_PASSWORDS, true)
+    )
+      return null
     check(saveFlags != null) { "saveFlags must not be null" }
     val idsToSave = scenario.fieldsToSave.toTypedArray()
     if (idsToSave.isEmpty()) return null
@@ -266,7 +277,27 @@ class Api30AutofillResponseBuilder private constructor(form: FillableForm) :
     if (scenario.hasUsername) {
       saveDataTypes = saveDataTypes or SaveInfo.SAVE_DATA_TYPE_USERNAME
     }
+
     return SaveInfo.Builder(saveDataTypes, idsToSave).run {
+
+      /**
+       * Suppress autofill save UI ("Save username and password to Password Store?") if at least one
+       * of .../<origin>/<username>.gpg or .../<origin>/<username>/password.gpg exists in the repo
+       */
+      val knownUsernames =
+        findKnownUsernamesFor(formOrigin.getPrettyIdentifier(context, untrusted = true))
+      if (
+        !knownUsernames.isEmpty() && scenario.username != null && scenario.username is AutofillId
+      ) {
+        val pattern =
+          Pattern.compile(
+            // negative (= not matching) regex
+            "^(?!(" + knownUsernames.map { Pattern.quote(it) }.joinToString("|") + ")$).*$"
+          )
+        val validator = RegexValidator(scenario.username as AutofillId, pattern)
+        setValidator(validator)
+      }
+
       setFlags(saveFlags)
       build()
     }